Move api folder to be in server
